pyproject.toml
[tool.poetry]
name = "zkp"
version = "0.1.0"
description = ""
authors = ["Jesus Cardenes"]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.4.1"
numpy = "^1.22.4"
pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
docker-compose = "^1.29.2"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.950"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
